requirements.txt
# Core
pyautogui>=0.9.54
pillow>=10.0.0

# Windows Accessibility (Windows only)
pywinauto>=0.6.8; sys_platform == 'win32'

# Web Automation (optional)
selenium>=4.15.0
webdriver-manager>=4.0.0
